支持自定义抛错样式

+++ core/Autumn.php
@@ -246,15 +246,30 @@ class Autumn
 	* 系统异常处理
 	* ======
 	* @author 洪波
-	* @version 16.03.30
+	* @version 17.01.10
 	*/
 	public function exception($content, $interrupt = true)
 	{
 		header("Content-Type:text/html; charset=utf-8");
-		echo '<div style="text-align:center;padding:10px;border:1px dashed #ccc;color:#ff4e00;background:#eee;">',
-			'<p style="color:#666;"><strong style="font-size:20px;">警告：系统异常</strong></p>',
-			'<div style="border-top:1px dashed #ccc; padding:20px;">',$content,'</div>',
-			'<p style="color:#999;"><small>Autumn版本：',self::FRAMEWORK_VERSION,'</small></p></div>';
+		$view = ROOT . '/app/views/exception.php';
+		if(is_file($view))
+		{
+			extract(array(
+				'error_detail' => $content,
+				'system_version' => self::FRAMEWORK_VERSION
+			), EXTR_PREFIX_SAME, 'data');
+			ob_start();
+			ob_implicit_flush(false);
+			require($view);
+			echo ob_get_clean();
+		}
+		else
+		{
+			echo '<div style="text-align:center;padding:10px;border:1px dashed #ccc;color:#ff4e00;background:#eee;">',
+				'<p style="color:#666;"><strong style="font-size:20px;">警告：系统异常</strong></p>',
+				'<div style="border-top:1px dashed #ccc; padding:20px;">',$content,'</div>',
+				'<p style="color:#999;"><small>Autumn版本：',self::FRAMEWORK_VERSION,'</small></p></div>';
+		}
 		if($interrupt)
 		{
 			exit;
